+++ crates/chainio/clients/avsregistry/src/writer.rs
@@ -7,6 +7,7 @@ use eigen_crypto_bls::{
     alloy_g1_point_to_g1_affine, convert_to_g1_point, convert_to_g2_point, BlsKeyPair,
 };
 use eigen_logging::logger::SharedLogger;
+use eigen_utils::slashing::middleware::registrycoordinator::ISlashingRegistryCoordinatorTypes::OperatorSetParam;
 use eigen_utils::slashing::middleware::registrycoordinator::{
     IBLSApkRegistryTypes::PubkeyRegistrationParams, ISignatureUtils::SignatureWithSaltAndExpiry,
     RegistryCoordinator,
@@ -360,7 +361,6 @@ impl AvsRegistryChainWriter {
         Ok(*tx.tx_hash())
     }
 
-    /// function ejectOperator(address operator, bytes memory quorumNumbers) external onlyEjector {
     pub async fn eject_operator(
         &self,
         operator_address: Address,
@@ -382,6 +382,36 @@ impl AvsRegistryChainWriter {
             .inspect(|tx| info!(tx_hash = ?tx, "successfully ejected operator from AVS"))
             .map(|tx| *tx.tx_hash())
     }
+
+    /// Update the configuration of an existing quorum.
+    ///
+    /// # Arguments
+    /// * `quorum_number` - the quorum number to update
+    /// * `operator_set_params` - the new config with [`OperatorSetParam`]
+    ///
+    /// # Returns
+    /// * `TxHash` - The transaction hash of the set operator set param transaction
+    pub async fn set_operator_set_param(
+        &self,
+        quorum_number: u8,
+        operator_set_params: OperatorSetParam,
+    ) -> Result<TxHash, AvsRegistryError> {
+        info!("setting operator set param with the AVS's registry coordinator");
+        let provider = get_signer(&self.signer.clone(), &self.provider);
+
+        let contract_registry_coordinator =
+            RegistryCoordinator::new(self.registry_coordinator_addr, provider);
+
+        let contract_call =
+            contract_registry_coordinator.setOperatorSetParams(quorum_number, operator_set_params);
+
+        let tx = contract_call
+            .send()
+            .await
+            .map_err(AvsRegistryError::AlloyContractError)?;
+        info!(tx_hash = ?tx,"successfully set operator set param with the AVS's registry coordinator" );
+        Ok(*tx.tx_hash())
+    }
 }
 
 #[cfg(test)]
@@ -401,6 +431,7 @@ mod tests {
     };
     use eigen_testing_utils::transaction::wait_transaction;
     use eigen_utils::rewardsv2::middleware::servicemanagerbase::ServiceManagerBase;
+    use eigen_utils::slashing::middleware::registrycoordinator::ISlashingRegistryCoordinatorTypes::OperatorSetParam;
     use eigen_utils::slashing::middleware::registrycoordinator::RegistryCoordinator;
     use futures_util::StreamExt;
     use std::str::FromStr;
@@ -598,7 +629,63 @@ mod tests {
         let mut stream = poller.into_stream();
         let (stream_event, _) = stream.next().await.unwrap().unwrap();
 
-        assert_eq!(stream_event.socket, new_socket_addr)
+        assert_eq!(stream_event.socket, new_socket_addr);
+    }
+
+    #[tokio::test]
+    async fn test_set_operator_set_param() {
+        let (_container, http_endpoint, _ws_endpoint) = start_m2_anvil_container().await;
+        let bls_key =
+            "1371012690269088913462269866874713266643928125698382731338806296762673180359922"
+                .to_string();
+        let private_key = ANVIL_FIRST_PRIVATE_KEY.to_string();
+        let avs_writer =
+            build_avs_registry_chain_writer(http_endpoint.clone(), private_key.clone()).await;
+        let quorum_nums = Bytes::from([0]);
+
+        test_register_operator(&avs_writer, bls_key, quorum_nums, http_endpoint.clone()).await;
+
+        let registry_coordinator_contract = RegistryCoordinator::new(
+            avs_writer.registry_coordinator_addr,
+            get_signer(&avs_writer.signer.clone(), &avs_writer.provider),
+        );
+
+        let operator_set_params = OperatorSetParam {
+            maxOperatorCount: 10,
+            kickBIPsOfOperatorStake: 50,
+            kickBIPsOfTotalStake: 50,
+        };
+
+        let tx_hash = avs_writer
+            .set_operator_set_param(0, operator_set_params.clone())
+            .await
+            .unwrap();
+
+        let tx_status = wait_transaction(&http_endpoint, tx_hash)
+            .await
+            .unwrap()
+            .status();
+
+        assert!(tx_status);
+
+        let op_params = registry_coordinator_contract
+            .getOperatorSetParams(0)
+            .call()
+            .await
+            .unwrap();
+
+        assert_eq!(
+            op_params._0.maxOperatorCount,
+            operator_set_params.maxOperatorCount
+        );
+        assert_eq!(
+            op_params._0.kickBIPsOfOperatorStake,
+            operator_set_params.kickBIPsOfOperatorStake
+        );
+        assert_eq!(
+            op_params._0.kickBIPsOfTotalStake,
+            operator_set_params.kickBIPsOfTotalStake
+        );
     }
 
     #[tokio::test]